// src/app/core/models/filtro.model.ts
export class FiltrosPedido {
  pagina: number;
  itensPorPagina: number;
  id: string;
  valor: string;
  nome: string;
  datapedidode: string;
  datapedidoate: string;
  statuspedido: string;
  transportadora: string;
  status: string;
  frete: string;
  plataforma: string;
  pedido: string;
}

export class FiltrosAtendimentos {
  pagina: number;
  itensPorPagina: number;
  id: string;
  idficha: string;
  nome: string;
  datanascde: string;
  datanascate: string;
  sexo: string;
  formapagamento: string;
  datalancamentode: string;
  datalancamentoate: string;
  emailusuario: string;
  status: string;
}